ui: refactor modals to be more modular, and use new setup flow's forms
