+++ ui/com/fab.jsx
@@ -1,6 +1,5 @@
 'use babel'
 import React from 'react'
-import { ModalBtn } from './modals'
 
 export default class FAB extends React.Component {
   render() {
