+++ ui/less/com/card.less
@@ -22,9 +22,9 @@
     color: #555;
     margin-bottom: 10px;
     border-radius: 5px;
-    text-align: center;
-    .btn {
+    &>span>.btn {
       display: block;
+      text-align: center;
       background: #fff;
       padding: 10px 20px;
       font-weight: bold;
