+++ ui/less/forms.less
@@ -46,7 +46,7 @@ form {
       text-align: center;
     }
     input {
-      width: 406px;
+      width: 100%;
       margin-bottom: 5px;
     }
   }
