+++ ui/less/views/sync.less
@@ -2,7 +2,7 @@
   background: #fafafa;
   .header {
     padding: 20px;
-    h1 {
+    &>h1 {
       font-weight: 100;
       font-size: 250%;
       margin: 0 0 12px;
@@ -14,7 +14,7 @@
       font-weight: 300;
       color: #555;
     }
-    .btn {
+    &>.btn {
       background: #fff;
     }
   }
